Moved the restoration logic into seperate module
