+++ src/aims_ui/static/js/a_single_matches/radiussearch/init.mjs
@@ -1,9 +1,10 @@
 import { setupSubmissionFromInputs } from "./form_helpers.mjs";
-import { updateMapToFitCircle, setupLatLongToMapListeners, addMatchedAddressMarkersToMap, setupMap, setupRadiusListeners, setupLatLongFromMapListeners, setupInitialMarkerLocation } from "./interactive_map/map_setup.mjs";
+import { updateMapToFitCircle, setupLatLongToMapListeners, setupMap, setupRadiusListeners, setupLatLongFromMapListeners, setupInitialMarkerLocation } from "./interactive_map/map_setup.mjs";
 import { setMapSizeAndResizeListeners } from "./interactive_map/map_size.mjs";
 import { setupHorizontalInputs } from "./page_reformatting.mjs";
 import { setupScrollListeners, setScrollPosition } from "./scroll_helpers.mjs";
 import { setupUprnSearchFunctionality } from "./uprn_helper.mjs";
+import { addMatchedAddressMarkersToMap } from "./interactive_map/result_address_markers.mjs";
 
 function setupMapSpecificItems() {
   // Setup the interactive map and return refference to it
@@ -22,7 +23,7 @@ function setupMapSpecificItems() {
   setupRadiusListeners();
 
   // Add any results that are sat in local storage onto the map!
-  addMatchedAddressMarkersToMap();
+  addMatchedAddressMarkersToMap(map);
 
   // Set map size and resize listeners
   setMapSizeAndResizeListeners(map);
